+++ ios/evy/Data/EVYDataManager.swift
@@ -13,32 +13,6 @@ public enum EVYDataError: Error {
     case keyNotFound
 }
 
-public enum EVYError: LocalizedError {
-    case parsingFailed(context: String)
-    case invalidData(context: String)
-    case regexCompilationFailed(pattern: String)
-    case imageLoadFailed(name: String)
-    case formatFailed(type: String, reason: String)
-    case websocketError(context: String)
-
-    public var errorDescription: String? {
-        switch self {
-        case .parsingFailed(let context):
-            return "Parsing failed: \(context)"
-        case .invalidData(let context):
-            return "Invalid data: \(context)"
-        case .regexCompilationFailed(let pattern):
-            return "Invalid regex pattern: \(pattern)"
-        case .imageLoadFailed(let name):
-            return "Failed to load image: \(name)"
-        case .formatFailed(let type, let reason):
-            return "Failed to format \(type): \(reason)"
-        case .websocketError(let context):
-            return "WebSocket error: \(context)"
-        }
-    }
-}
-
 extension Notification.Name {
     static let evyDataUpdated = Notification.Name("EVYDataUpdated")
     static let evyFlowUpdated = Notification.Name("EVYFlowUpdated")
